Migrate refund payment handler integration test to e2e (#1081)
* Cancel payment integration test migration to e2e.

* Refund payment integration test migration to e2e.

* Fix eslint issue.

* Debug klarna payment e2e test

* Fix package.json

* Change e2e test name

* Fix format

* Fix flaky payapal e2e test

---------
