+++ extension/test/e2e/pageObjects/PaypalPopupPage.js
@@ -10,10 +10,11 @@ export default class PaypalPopupPage {
     await this.page.type('#email', 'sb-urfdg23058190@personal.example.com')
     await this.page.click('#btnNext')
 
-    await this.page.waitForTimeout(1000)
+    await this.page.waitForTimeout(4_000)
 
-    await this.page.click('#acceptAllButton')
+    // await this.page.click('#acceptAllButton')
     await this.page.type('#password', '!nh-NNS1')
+    await this.page.waitForSelector('#btnLogin')
     await this.page.click('#btnLogin')
 
     await this.page.waitForSelector('#payment-submit-btn')
